mail_serviceの処理の見直し

+++ bin/dipper.sh
@@ -13,42 +13,31 @@ if [ -e ${User_File} ]; then
     source "${User_File}"
 fi
 
-# エラーメール通知機能がonの場合は初期化処理を行い。offの場合はそれぞれのキャッシュファイルを削除する
 mail_service() {
-    # キャッシュファイルのパス
     local cache_dir="../cache"
     local cache_err="${cache_dir}/err_mail"
     local cache_ddns="${cache_dir}/ddns_mail"
 
     if [[ -n ${EMAIL_CHK_ADR:-} ]]; then
         if [[ -n ${ERR_CHK_TIME:-} ]]; then
-            # err_mail_cache初期化
-            ./mail_handle.sh "err_mail" "dipperでエラーを検出しました" "$EMAIL_CHK_ADR" &      
-            ./err_mail_service.sh "$EMAIL_CHK_ADR" "$ERR_CHK_TIME" &    # err_mail_serviceのタイマーを起動
-        else
-            if [ -f "${cache_err}" ]; then
-                rm "${cache_err}"
-            fi
-            if [ ! -f "${cache_ddns}" ] && [ -d "${cache_dir}" ]; then
-                rm -r "${cache_dir}"
-            fi
+            ./mail_handle.sh "err_mail" "dipperでエラーを検出しました" "$EMAIL_CHK_ADR" &
+            ./err_mail_service.sh "$EMAIL_CHK_ADR" "$ERR_CHK_TIME" &
+        elif [ -f "${cache_err}" ]; then
+            rm "${cache_err}"
         fi
 
         if [[ -n ${EMAIL_CHK_DDNS:-} ]]; then
-            # ddns_mail_cache初期化
-            ./mail_handle.sh "ddns_mail" "IPアドレスの変更がありました" "$EMAIL_CHK_ADR" & 
-        else
-            if [ -f "${cache_ddns}" ]; then
-                rm "${cache_ddns}"
-            fi
-            if [ ! -f "${cache_err}" ] && [ -d "${cache_dir}" ]; then
-                rm -r "${cache_dir}"
-            fi
+            ./mail_handle.sh "ddns_mail" "IPアドレスの変更がありました" "$EMAIL_CHK_ADR" &
+        elif [ -f "${cache_ddns}" ]; then
+            rm "${cache_ddns}"
         fi
-    else
-        if [ -d "${cache_dir}" ]; then
+        # ディレクトリの中身をチェック
+        if [ -d "${cache_dir}" ] && [ -z "$(ls -A ${cache_dir})" ]; then
+            # ファイルが存在しない
             rm -r "${cache_dir}"
         fi
+    elif [ -d "${cache_dir}" ]; then
+        rm -r "${cache_dir}"
     fi
 }
